requirements.txt
requests
python-dotenv
pydantic
loguru
openai